postgre conf, cleaned up comments, admin config, testcases, added requirements
