+++ .gitignore
@@ -105,3 +105,7 @@ venv.bak/
 
 # sqlite db
 *.sqlite3
+
+# Replays folder
+Replays/
+!Replays/.gitkeep
