+++ Overmind/settings.py
@@ -1,18 +1,6 @@
-"""
-Django settings for Overmind project.
-
-Generated by 'django-admin startproject' using Django 2.1.1.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/2.1/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/2.1/ref/settings/
-"""
-
 import os
 
-# Define where to store replay files
+# Define folder where to store replay files
 REPLAYS_DIR = '/home/sc2/dev/Overmind/Replays/'
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
@@ -77,8 +65,12 @@ WSGI_APPLICATION = 'Overmind.wsgi.application'
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'NAME': 'overmind',
+        'USER': 'overmind',
+        'PASSWORD': 'overmind',
+        'HOST': 'localhost',
+        'PORT': '',
     }
 }
