--- Overmind/settings/dev_sample.py
@@ -1,8 +1,0 @@
-from Overmind.settings.base import *
-
-DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': os.path.join(BASE_DIR, 'dev_db.sqlite3'),
-    }
-}
