+++ Overmind/wsgi.py
@@ -1,12 +1,3 @@
-"""
-WSGI config for Overmind project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/2.1/howto/deployment/wsgi/
-"""
-
 import os
 
 from django.core.wsgi import get_wsgi_application
