+++ requirements.txt
@@ -1,0 +1,3 @@
+django==2.1
+tastypie
+psycopg2
